fixed course, tags

+++ src/crud/tag.py
@@ -16,6 +16,12 @@ def get_tag_by_id(db: Session, tag_id: UUID):
     return tag
 
 def create_tags(db: Session, payload: CreateTag):
+    name = payload.name.strip()
+    if not name:
+        raise HTTPException(status_code=400, detail="Tag name cannot be empty")
+    existing_tag = db.query(TagModel).filter(TagModel.name == name).first()
+    if existing_tag:
+        raise HTTPException(status_code=400, detail="Tag already exists")
     tag = TagModel(name=payload.name)
     db.add(tag)
     db.commit()
@@ -24,13 +30,15 @@ def create_tags(db: Session, payload: CreateTag):
 
 def delete_tags(db: Session, tag_id: UUID):
     tag = db.query(TagModel).filter(TagModel.id == tag_id).first()
+    if tag is None:
+        raise HTTPException(status_code=404, detail="Tag not found")
     db.delete(tag)
     db.commit()
     return {"message": f"{tag.name} is deleted"}
 
-def add_tag_to_course(db: Session, payload: CourseTag):
-    course = get_course_by_id(db, payload.course_id)
-    tag = get_tag_by_id(db, payload.tag_id)
+def add_tag_to_course(db: Session, course_id: UUID, tag_id: UUID):
+    course = get_course_by_id(db, course_id)
+    tag = get_tag_by_id(db, tag_id)
     if course is None:
         raise HTTPException(status_code=404, detail="Course not found")
     if tag is None:
